Format sol files with prettier

+++ contracts/BrewlabsPriceOracle.sol
@@ -5,6 +5,7 @@ import "@openzeppelin/contracts/utils/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
 import "./libs/PriceOracle.sol";
+
 interface AggregatorV3Interface {
     function decimals() external view returns (uint8);
     function description() external view returns (string memory);
@@ -13,21 +14,15 @@ interface AggregatorV3Interface {
     // getRoundData and latestRoundData should both raise "No data present"
     // if they do not have data to report, instead of returning unset values
     // which could be misinterpreted as actual reported values.
-    function getRoundData(uint80 _roundId) external view returns (
-        uint80 roundId,
-        int256 answer,
-        uint256 startedAt,
-        uint256 updatedAt,
-        uint80 answeredInRound
-    );
+    function getRoundData(uint80 _roundId)
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
 
-    function latestRoundData() external view returns (
-        uint80 roundId,
-        int256 answer,
-        uint256 startedAt,
-        uint256 updatedAt,
-        uint80 answeredInRound
-    );
+    function latestRoundData()
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
 }
 
 interface IERC20Extended is IERC20 {
@@ -59,26 +54,26 @@ contract BrewlabsPriceOracle is PriceOracle {
     address public wrapped;
 
     /// @notice Chainlink Aggregators
-    mapping(address => AggregatorV3Interface) public aggregators;    
+    mapping(address => AggregatorV3Interface) public aggregators;
 
     struct PriceInfo {
-        address token;              // Address of token contract, TOKEN
-        address baseToken;          // Address of base token contract, BASETOKEN
-        address lpToken;            // Address of TOKEN-BASETOKEN pair contract
-        bool active;                // Active status of price record 0 
+        address token; // Address of token contract, TOKEN
+        address baseToken; // Address of base token contract, BASETOKEN
+        address lpToken; // Address of TOKEN-BASETOKEN pair contract
+        bool active; // Active status of price record 0
     }
 
     struct TwapPriceInfo {
-        address token;              // Address of token contract, TOKEN
-        address baseToken;          // Address of base token contract, BASETOKEN
-        address twapOracle;         // Address of twap oracle contract
-        bool active;                // Active status of price record 0 
+        address token; // Address of token contract, TOKEN
+        address baseToken; // Address of base token contract, BASETOKEN
+        address twapOracle; // Address of twap oracle contract
+        bool active; // Active status of price record 0
     }
 
     mapping(address => PriceInfo) public priceRecords;
     mapping(address => TwapPriceInfo) public priceTwapRecords;
     mapping(address => uint256) public assetPrices;
-    
+
     event NewAdmin(address oldAdmin, address newAdmin);
     event PriceRecordUpdated(address token, address baseToken, address lpToken, bool _active);
     event TwapPriceRecordUpdated(address token, address baseToken, address twapOracle, bool _active);
@@ -104,7 +99,7 @@ contract BrewlabsPriceOracle is PriceOracle {
         }
         if (tokenPrice == 0) {
             tokenPrice = getPriceFromDex(tokenAddress);
-        } 
+        }
         return tokenPrice;
     }
 
@@ -113,10 +108,10 @@ contract BrewlabsPriceOracle is PriceOracle {
         if (priceInfo.active) {
             uint256 rawTokenAmount = IERC20Extended(priceInfo.token).balanceOf(priceInfo.lpToken);
             uint256 tokenDecimalDelta = 18 - uint256(IERC20Extended(priceInfo.token).decimals());
-            uint256 tokenAmount = rawTokenAmount.mul(10**tokenDecimalDelta);
+            uint256 tokenAmount = rawTokenAmount.mul(10 ** tokenDecimalDelta);
             uint256 rawBaseTokenAmount = IERC20Extended(priceInfo.baseToken).balanceOf(priceInfo.lpToken);
             uint256 baseTokenDecimalDelta = 18 - uint256(IERC20Extended(priceInfo.baseToken).decimals());
-            uint256 baseTokenAmount = rawBaseTokenAmount.mul(10**baseTokenDecimalDelta);
+            uint256 baseTokenAmount = rawBaseTokenAmount.mul(10 ** baseTokenDecimalDelta);
             uint256 baseTokenPrice = getPriceFromOracle(priceInfo.baseToken);
             uint256 tokenPrice = baseTokenPrice.mul(baseTokenAmount).div(tokenAmount);
 
@@ -129,9 +124,12 @@ contract BrewlabsPriceOracle is PriceOracle {
     function getPriceFromTwap(address _tokenAddress) public view returns (uint256) {
         TwapPriceInfo storage priceInfo = priceTwapRecords[_tokenAddress];
         if (priceInfo.active) {
-            uint144 twapPrice = ITwapOracle(priceInfo.twapOracle).twap(priceInfo.token, 10**(uint256(IERC20Extended(priceInfo.token).decimals())));
+            uint144 twapPrice = ITwapOracle(priceInfo.twapOracle).twap(
+                priceInfo.token, 10 ** (uint256(IERC20Extended(priceInfo.token).decimals()))
+            );
             uint256 baseTokenPrice = getPriceFromOracle(priceInfo.baseToken);
-            uint256 tokenPrice = baseTokenPrice.mul(twapPrice).div(10**(uint256(IERC20Extended(priceInfo.token).decimals())));
+            uint256 tokenPrice =
+                baseTokenPrice.mul(twapPrice).div(10 ** (uint256(IERC20Extended(priceInfo.token).decimals())));
             return tokenPrice;
         } else {
             return 0;
@@ -146,7 +144,7 @@ contract BrewlabsPriceOracle is PriceOracle {
     function getPriceFromChainlink(address _tokenAddress) public view returns (uint256) {
         AggregatorV3Interface aggregator = aggregators[_tokenAddress];
         if (address(aggregator) != address(0)) {
-            ( , int answer, , , ) = aggregator.latestRoundData();
+            (, int256 answer,,,) = aggregator.latestRoundData();
 
             // It's fine for price to be 0. We have two price feeds.
             if (answer == 0) {
@@ -154,12 +152,12 @@ contract BrewlabsPriceOracle is PriceOracle {
             }
 
             // Extend the decimals to 1e18.
-            uint retVal = uint(answer);
-            uint price = retVal.mul(10**(18 - uint(aggregator.decimals())));
+            uint256 retVal = uint256(answer);
+            uint256 price = retVal.mul(10 ** (18 - uint256(aggregator.decimals())));
 
-            return price;            
+            return price;
         }
-        return 0;        
+        return 0;
     }
 
     function setDexPriceInfo(address _token, address _baseToken, address _lpToken, bool _active) external {
@@ -202,11 +200,11 @@ contract BrewlabsPriceOracle is PriceOracle {
 
     function setAggregators(address[] calldata tokenAddresses, address[] calldata sources) external {
         require(msg.sender == admin, "only the admin may set the aggregators");
-        for (uint i = 0; i < tokenAddresses.length; i++) {
+        for (uint256 i = 0; i < tokenAddresses.length; i++) {
             aggregators[tokenAddresses[i]] = AggregatorV3Interface(sources[i]);
             emit AggregatorUpdated(tokenAddresses[i], sources[i]);
         }
-    } 
+    }
 
     function compareStrings(string memory a, string memory b) internal pure returns (bool) {
         return (keccak256(abi.encodePacked((a))) == keccak256(abi.encodePacked((b))));
