+++ contracts/others/BaltoTeamLocker.sol
@@ -6,10 +6,9 @@ pragma solidity ^0.8.0;
  * This contract has been developed by brewlabs.info
  */
 
-import '@openzeppelin/contracts/access/Ownable.sol';
-import '@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol';
-import '@openzeppelin/contracts/utils/math/SafeMath.sol';
-
+import "@openzeppelin/contracts/access/Ownable.sol";
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import "@openzeppelin/contracts/utils/math/SafeMath.sol";
 
 contract BaltoTeamLocker is Ownable {
     using SafeMath for uint256;
@@ -23,7 +22,7 @@ contract BaltoTeamLocker is Ownable {
 
     uint256 public dividendPercent = 9800;
     uint256 private PRECISION_FACTOR = 1 ether;
-   
+
     mapping(address => bool) isUsed;
     mapping(address => bool) public isDistributor;
     address[] public distributors;
@@ -36,13 +35,13 @@ contract BaltoTeamLocker is Ownable {
     event Harvested(uint256 amount);
     event EmergencyWithdrawn();
     event EmergencyDividendWithdrawn();
-        
+
     modifier onlyActive() {
         require(isActive == true, "not active");
         _;
     }
 
-    constructor () {}
+    constructor() {}
 
     function initialize(IERC20 _token, address _reflectionToken) external onlyOwner {
         require(initialized == false, "already initialized");
@@ -52,12 +51,11 @@ contract BaltoTeamLocker is Ownable {
         reflectionToken = _reflectionToken;
     }
 
-
     function addDistribution(address distributor) external onlyOwner {
         require(isDistributor[distributor] == false, "already set");
 
         isDistributor[distributor] = true;
-        if(!isUsed[distributor]) {
+        if (!isUsed[distributor]) {
             distributors.push(distributor);
             isUsed[distributor] = true;
         }
@@ -71,7 +69,7 @@ contract BaltoTeamLocker is Ownable {
 
         isDistributor[distributor] = false;
         totalDistributors = totalDistributors.sub(1);
-        
+
         emit RemoveDistribution(distributor);
     }
 
@@ -79,20 +77,20 @@ contract BaltoTeamLocker is Ownable {
         require(isDistributor[msg.sender] == true || msg.sender == owner(), "only distributor");
 
         uint256 reflectionTokens = 0;
-        if(reflectionToken == address(0x0)) {
+        if (reflectionToken == address(0x0)) {
             reflectionTokens = address(this).balance;
         } else {
             reflectionTokens = IERC20(reflectionToken).balanceOf(address(this));
         }
 
         uint256 dAmt = reflectionTokens.mul(dividendPercent).div(10000).div(totalDistributors);
-        if(dAmt == 0) return;
+        if (dAmt == 0) return;
 
-        for(uint i = 0; i < distributors.length; i++) {
+        for (uint256 i = 0; i < distributors.length; i++) {
             address distributor = distributors[i];
-            if(!isDistributor[distributor]) continue;
+            if (!isDistributor[distributor]) continue;
 
-            if(reflectionToken == address(0x0)) {
+            if (reflectionToken == address(0x0)) {
                 payable(distributor).transfer(dAmt);
             } else {
                 IERC20(reflectionToken).safeTransfer(distributor, dAmt);
@@ -103,10 +101,10 @@ contract BaltoTeamLocker is Ownable {
     }
 
     function pendingReflection(address _user) external view returns (uint256) {
-        if(isDistributor[_user] == false) return 0;
+        if (isDistributor[_user] == false) return 0;
 
         uint256 reflectionTokens = 0;
-        if(reflectionToken == address(0x0)) {
+        if (reflectionToken == address(0x0)) {
             reflectionTokens = address(this).balance;
         } else {
             reflectionTokens = IERC20(reflectionToken).balanceOf(address(this));
@@ -127,14 +125,14 @@ contract BaltoTeamLocker is Ownable {
 
     function emergencyWithdraw() external onlyOwner {
         uint256 tokenAmt = token.balanceOf(address(this));
-        if(tokenAmt > 0) {
+        if (tokenAmt > 0) {
             token.transfer(msg.sender, tokenAmt);
         }
         emit EmergencyWithdrawn();
     }
 
     function emergencyDividendWithdraw() external onlyOwner {
-        if(reflectionToken == address(0x0)) {
+        if (reflectionToken == address(0x0)) {
             uint256 reflectionTokens = address(this).balance;
             payable(msg.sender).transfer(reflectionTokens);
         } else {
@@ -144,14 +142,14 @@ contract BaltoTeamLocker is Ownable {
 
         emit EmergencyDividendWithdrawn();
     }
-    
+
     /**
      * @notice It allows the admin to recover wrong tokens sent to the contract
      * @param _token: the address of the token to withdraw
      * @dev This function is only callable by admin.
      */
     function rescueToken(address _token) external onlyOwner {
-        if(_token == address(0x0)) {
+        if (_token == address(0x0)) {
             uint256 _tokenAmount = address(this).balance;
             payable(msg.sender).transfer(_tokenAmount);
         } else {
